requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
gTTS==2.5.1
googletrans-py==4.0.0
aiosqlite==0.19.0